requirements.txt
pywin32==308
pycryptodome==3.21.0
PyQt5==5.15.11